2007-06-20  Murray Cumming  <murrayc@murrayc.com>
* src/modest-ui-actions.c: (modest_ui_actions_on_details):
Add a g_warning when the item is not a folder.

* src/modest-ui-dimming-rules.c:
(modest_ui_dimming_rules_on_details): Dim the Email/Details menu item
when selected item is not a folder (for instance, an account), because
the UI spec only talks about folders.
I am not sure what the UI spec wants, so this might not totally fix
projects.maemo.org bug NB#60794.

pmo-trunk-r2342