2007-07-16  Murray Cumming  <murrayc@murrayc.com>
Reverted this previous commit from Johannes,
because the scrollbars are always visible and because
I think we only need it on onepage of the account
settings, not the account wizard.
(I wrote the wrong instructions in the bug.)

* src/maemo/easysetup/modest-easysetup-wizard.c:
Added scrollbars to account wizard dialog.
Fixed projects.maemo.org NB #62744

pmo-trunk-r2765