2007-07-30  Murray Cumming  <murrayc@murrayc.com>
* src/modest-tny-account-store.c: (on_account_removed),
(on_account_changed): Emit the ACCOUNT_UPDATE signal so that the
account is removed from the View menu. We have TnyAccountStore
and ModestTnyAccountStore signals that seem very similar, but this
seems to work for now.

* src/maemo/modest-main-window.c:
(on_show_account_action_activated): Prevent dereference
of a NULL if the accoutn does not really exist, though that
should not happen anymore.

This fixes projects.maemo.org bug NB#64606.

pmo-trunk-r2855