2007-08-13  Murray Cumming  <murrayc@murrayc.com>
* src/dbus_api/modest-dbus-callbacks.c: (modest_dbus_req_filter):
Comment out the debug warning about the unexpected D-Bus methods.
They are just annoying now.
* src/maemo/modest-platform.c:
(modest_platform_set_update_interval): Correct the D-Bus path, instead
of using a file path, though it did not seem to break things before.
It might be needed now:
Use the ALARM_EVENT_ACTIVATION flag so that emails are updated even when
modest is not already running. This is the behaviour of the old client, and
Mox says it is what we should do for modest.
This closes projects.maemo.org bug NB#64327.

pmo-trunk-r2965