2007-09-11  Murray Cumming  <murrayc@murrayc.com>
* src/maemo/modest-account-settings-dialog.c:
(modest_account_settings_dialog_set_account_name): Set the incoming
secure-auth checkbox only when a secure auth method is really specified
in gconf. For some reason, it was set also if a secure connection was
chosen, but that does not make sense to me, and was not commented.
For instance, this stops the user from seeing an error about unsupported
secure authentication when editing a gmail account.

pmo-trunk-r3274